+++ p2panda-spaces/src/args.rs
@@ -1,5 +1,2 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-pub mod dgm;
-pub mod message;
-pub mod orderer;
