+++ p2panda-spaces/src/auth/mod.rs
@@ -1,3 +1,4 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+pub mod message;
 pub mod orderer;
