+++ p2panda-spaces/src/encryption/orderer.rs
@@ -8,6 +8,7 @@ use p2panda_encryption::data_scheme::GroupSecretId;
 use p2panda_encryption::traits::GroupMessage as EncryptionOperation;
 
 use crate::encryption::dgm::EncryptionGroupMembership;
+use crate::encryption::message::{EncryptionArgs, EncryptionMessage};
 use crate::types::{ActorId, EncryptionControlMessage, EncryptionDirectMessage, OperationId};
 
 #[derive(Clone, Debug)]
@@ -106,25 +107,4 @@ impl<M> EncryptionOperation<ActorId, OperationId, EncryptionGroupMembership>
     fn direct_messages(&self) -> Vec<EncryptionDirectMessage> {
         todo!()
     }
-}
-
-#[derive(Clone, Debug)]
-pub enum EncryptionArgs {
-    // @TODO: Here we will fill in the "dependencies", which will be later used by ForgeArgs.
-    System {
-        control_message: EncryptionControlMessage,
-        direct_messages: Vec<EncryptionDirectMessage>,
-    },
-    Application {
-        group_secret_id: GroupSecretId,
-        nonce: XAeadNonce,
-        ciphertext: Vec<u8>,
-    },
-}
-
-#[derive(Clone, Debug)]
-#[allow(clippy::large_enum_variant)]
-pub enum EncryptionMessage<M> {
-    Args(EncryptionArgs),
-    Forged(M),
 }
